move set related models to their own file
