+++ components/session/records/sets/RenderSetField.tsx
@@ -1,12 +1,13 @@
 import { Box, Stack, TextFieldProps } from '@mui/material'
 import { memo } from 'react'
+import { TIME_FORMAT } from '../../../../lib/frontend/constants'
 import { UpdateFields } from '../../../../lib/util'
 import { VisibleField } from '../../../../models/DisplayFields'
-import { DB_UNITS, Set, Units, convertUnit } from '../../../../models/Set'
+import { Set } from '../../../../models/Set'
+import { DB_UNITS, Units, convertUnit } from '../../../../models/Units'
 import NumericFieldAutosave from '../../../form-fields/NumericFieldAutosave'
 import SetFieldSide from './SetFieldSide'
 import SetFieldTimeMask from './SetFieldTimeMask'
-import { TIME_FORMAT } from '../../../../lib/frontend/constants'
 
 const delimiterWidth = '15px'
 type ComponentType = 'side' | 'time' | 'default'
