+++ models/Units.test.ts
@@ -1,4 +1,4 @@
-import { convertUnit } from './Set'
+import { convertUnit } from './Units'
 
 describe('convertUnit', () => {
   it('converts effort', () => {
