+++ scripts/mongoDevDataGenerator.ts
@@ -15,8 +15,8 @@ import {
 import { createNote } from '../models/Note'
 import { createRecord } from '../models/Record'
 import { createSessionLog } from '../models/SessionLog'
-import { DB_UNITS, Units } from '../models/Set'
 import { Status } from '../models/Status'
+import { DB_UNITS, Units } from '../models/Units'
 import './polyfills'
 
 const envPath = path.resolve(__dirname, '..', '.env.development')
